move thkids test over to other test file
